Adding a checkbox for js vs json file type export
I'm not sure what the history of this project is, but fundamentally it's
exporting JSON, and there's no typeface.js popular library as far as I
can tell. It seems like there's some overlap with Three.js but I don't
know what the overlap is.

This adds a checkbox to export JSON and strip the typeface JS header. See
https://github.com/mrdoob/three.js/issues/8262#issuecomment-191115072
for some more relevant details. Three.js currently expects the header
text to never change and be exactly 65 characters in length. This is
also problematic because fundamentally this is just a JSON file, and no
typeface library is loaded. A more robust solution is to just export raw
JSON, and then downstream Three.js can consume the JSON as is.